Refactor EventService, add new ResultService
This commit refactors the EventService by extracting the logic related to result calculation into a new service, ResultService. This separation simplifies the code structure, improves modularity and makes the responsibility areas of components explicit. EventService now focuses on managing the life-cycle of an event, while ResultService concentrates on calculating and saving event results, both for individual matches and total scores.